subject crud functionality added

+++ public/assets/js/myapp.js
@@ -34,7 +34,7 @@ app.config(function ($routeProvider) {
   });
   $routeProvider.when("/subject", {
     templateUrl: "subject.html",
-    controller: "mainctrl",
+    controller: "subjectctrl",
     resolve: {
       check: function ($location, $rootScope) {
         if (!IsLoggedIn()) {
@@ -107,6 +107,147 @@ app.controller("loginctrl", function ($scope, $rootScope, $location, $http) {
   $scope.register = function () {
     $location.path("/registration");
   };
+});
+
+app.controller("subjectctrl", function ($scope, $http, $location, $rootScope) {
+  $scope.subjectList = [];
+  const initDate = () => {
+    const date = new Date();
+    date.setHours(7);
+    date.setMinutes(30);
+    date.setSeconds(0);
+    date.setMilliseconds(0);
+    return date;
+  };
+  $scope.newSubject = {
+    time_1: initDate(),
+    time_2: initDate(),
+  };
+
+  $scope.RetrieveSubjects = () => {
+    $http({
+      method: "get",
+      url: "subject",
+    }).then((response) => {
+      $scope.subjectList = response.data.subjects;
+    });
+  };
+
+  // Toggle Edit Subject
+  $scope.ToggleEdit = (subject) => {
+    if (!subject.time_1 && !subject.time_2) {
+      const time_1_orig = subject.time.split("-");
+      subject.time_1 = new Date();
+      subject.time_1.setHours(time_1_orig[0].split(":")[0]);
+      subject.time_1.setMinutes(time_1_orig[0].split(":")[1]);
+      subject.time_1.setSeconds(0);
+      subject.time_1.setMilliseconds(0);
+
+      subject.time_2 = new Date();
+      subject.time_2.setHours(time_1_orig[1].split(":")[0]);
+      subject.time_2.setMinutes(time_1_orig[1].split(":")[1]);
+      subject.time_2.setSeconds(0);
+      subject.time_2.setMilliseconds(0);
+    } else {
+      const time_1 = new Date(subject.time_1);
+      const time_2 = new Date(subject.time_2);
+      subject.time =
+        (time_1.getHours() > 12 ? time_1.getHours() - 12 : time_1.getHours()) +
+        ":" +
+        time_1.getMinutes() +
+        "-" +
+        (time_2.getHours() > 12 ? time_2.getHours() - 12 : time_2.getHours()) +
+        ":" +
+        time_2.getMinutes();
+    }
+
+    subject.edit = !subject.edit;
+    subject.editted = true;
+  };
+
+  $scope.DeleteSubject = (edpcode) => {
+    const deleteConfirmation = prompt("Type 'delete' to proceed");
+    if (deleteConfirmation.toLocaleLowerCase() !== "delete") return;
+    $http({
+      method: "delete",
+      url: `subject/${edpcode}`,
+    }).then((r) => {
+      $scope.RetrieveSubjects();
+      alert("Subject deleted");
+    });
+  };
+
+  $scope.SaveTable = () => {
+    let rowsAffected = 0;
+    for (let subject of $scope.subjectList) {
+      if (subject.editted) {
+        const time_1 = new Date(subject.time_1);
+        const time_2 = new Date(subject.time_2);
+        subject.time =
+          (time_1.getHours() > 12
+            ? time_1.getHours() - 12
+            : time_1.getHours()) +
+          ":" +
+          time_1.getMinutes() +
+          "-" +
+          (time_2.getHours() > 12
+            ? time_2.getHours() - 12
+            : time_2.getHours()) +
+          ":" +
+          time_2.getMinutes();
+
+        $http({
+          method: "put",
+          url: "subject",
+          data: subject,
+        });
+
+        rowsAffected++;
+      }
+
+      alert(`Data saved! ${rowsAffected} row(s) affected.`);
+
+      $scope.RetrieveSubjects();
+    }
+  };
+
+  $scope.SaveSubject = (newSubject) => {
+    if (!newSubject.edpcode) {
+      alert("Edpcode should not be empty!");
+      return;
+    }
+    let subject = {
+      edpcode: newSubject.edpcode ?? "",
+      subjectcode: newSubject.subjectcode ?? "",
+      days: `${newSubject.M ? "M" : ""}${newSubject.T ? "T" : ""}${
+        newSubject.W ? "W" : ""
+      }${newSubject.TH ? "TH" : ""}${newSubject.F ? "F" : ""}${
+        newSubject.S ? "S" : ""
+      }`,
+      room: newSubject.room ?? "",
+    };
+
+    const time_1 = new Date(newSubject.time_1);
+    const time_2 = new Date(newSubject.time_2);
+    subject.time =
+      (time_1.getHours() > 12 ? time_1.getHours() - 12 : time_1.getHours()) +
+      ":" +
+      time_1.getMinutes() +
+      "-" +
+      (time_2.getHours() > 12 ? time_2.getHours() - 12 : time_2.getHours()) +
+      ":" +
+      time_2.getMinutes();
+    $http({
+      method: "post",
+      url: "subject",
+      data: subject,
+    }).then((r) => {
+      $scope.RetrieveSubjects();
+      alert("Subject added.");
+    });
+  };
+
+  $scope.RetrieveSubjects();
 });
 ///
 app.controller("mainctrl", function ($scope, $http, $location, $rootScope) {
@@ -184,6 +325,8 @@ app.controller("mainctrl", function ($scope, $http, $location, $rootScope) {
 
   // delete student
   $scope.DeleteStudent = (idno) => {
+    const deleteConfirmation = prompt("Type 'delete' to proceed");
+    if (deleteConfirmation.toLocaleLowerCase() !== "delete") return;
     $http({
       method: "delete",
       url: `students/${idno}`,
